Add logging to support tuning auto paths
Log the distance traveled during a tuning auto path. This is useful to compare against the physically measured distance on the carpet.
Log the difference between the final pose based on odometry and the final target pose. This is useful to analyze how close the robot thinks it is getting to the final pose in an auto.