Log the vision pose estimate error.
Maintain an buffer of poses such that we can interpolate the pose corresponding to the timestamp of the vision pose estimate.
Log the difference between the vision pose estimate and this interpolated pose estimate.
This is useful for validating the positions of AprilTags on the field when the roobt is not moving and for tuning the vision pipeline latency when the robot is moving.